requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
anyio=3.2.1=pypi_0
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py38h06a4308_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
glib=2.68.2=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pypi_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py38_0
jinja2=3.0.1=pypi_0
jpeg=9b=h024ee3a_2
json5=0.9.6=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-server=1.9.0=pypi_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab=3.0.16=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=2.6.0=pypi_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py38h06a4308_0
matplotlib-base=3.3.4=py38h62a2d02_0
mistune=0.8.4=pypi_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
nbclassic=0.3.1=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.0=pypi_0
numpy=1.20.2=py38h2d18471_0
numpy-base=1.20.2=py38hfae3a4d_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pypi_0
pandas=1.2.5=py38h295c915_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py38he98fc37_0
pip=21.1.2=py38h06a4308_0
prometheus-client=0.11.0=pypi_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=pypi_0
pygments=2.9.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.9.0=py3.8_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=20.0.0=py38h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
requests-unixsocket=0.2.0=pypi_0
send2trash=1.7.1=pypi_0
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=pypi_0
sqlite=3.36.0=hc218d9a_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu102
tornado=6.1=py38h27cfd23_0
tqdm=4.61.1=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
websocket-client=1.1.0=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0

dataclasses~=0.8
scikit-image~=0.18.1
tqdm~=4.61.1
torchvision~=0.10.0
matplotlib~=3.3.4
numpy~=1.20.1
pillow~=8.3.1